More code removal from activity, created a separated VM for User Info
