+++ androidApp/src/main/java/com/example/emobilitychargingstations/android/MainActivity.kt
@@ -1,48 +1,27 @@
 package com.example.emobilitychargingstations.android
 
-import android.location.Location
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.*
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
 import com.comsystoreply.emobilitychargingstations.android.BuildConfig
 import com.comsystoreply.emobilitychargingstations.android.MyApplicationTheme
-import com.example.emobilitychargingstations.android.ui.composables.NavigationHostComposable
 import com.example.emobilitychargingstations.android.ui.composables.CarConnectionComposable
+import com.example.emobilitychargingstations.android.ui.composables.NavigationHostComposable
 import com.example.emobilitychargingstations.android.ui.utilities.LocationRequestStarter
-import com.example.emobilitychargingstations.models.UserLocation
-import com.google.android.gms.location.LocationCallback
-import com.google.android.gms.location.LocationResult
+import com.example.emobilitychargingstations.android.ui.viewmodels.StationsViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.osmdroid.config.Configuration
 
 class MainActivity : ComponentActivity() {
 
     private val stationsViewModel: StationsViewModel by viewModel()
-    private val locationCallback = object : LocationCallback() {
-        override fun onLocationResult(locationResult: LocationResult) {
-            locationResult.locations.firstOrNull()?.let {
-                if (checkIsDebugLocationMocked(it))  {
-                    stationsViewModel.setUserLocation(
-                        UserLocation(
-                            it.latitude,
-                            it.longitude
-                        )
-                    )
-                    stationsViewModel.startRepeatingStationsRequest()
-                }
-            }
-        }
-    }
-
-    private fun checkIsDebugLocationMocked(location: Location) : Boolean {
-        return if (BuildConfig.DEBUG) location.isMock else true
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -79,7 +58,7 @@ class MainActivity : ComponentActivity() {
     }
 
     private fun startRequestingLocation() {
-        LocationRequestStarter(this, locationCallback)
+        LocationRequestStarter(this, stationsViewModel.locationCallback)
     }
 
     companion object {
