+++ androidApp/src/main/java/com/example/emobilitychargingstations/android/di/AndroidKoinModule.kt
@@ -2,7 +2,8 @@ package com.example.emobilitychargingstations.android.di
 
 import android.app.Application
 import app.cash.sqldelight.db.SqlDriver
-import com.example.emobilitychargingstations.android.StationsViewModel
+import com.example.emobilitychargingstations.android.ui.viewmodels.StationsViewModel
+import com.example.emobilitychargingstations.android.ui.viewmodels.UserViewModel
 import com.example.emobilitychargingstations.data.local.DatabaseDriverFactory
 import org.koin.dsl.module
 
@@ -13,6 +14,7 @@ fun androidKoinModule() = module {
 
 fun viewModelModule() = module {
     factory { StationsViewModel(get(), get()) }
+    factory { UserViewModel(get()) }
 }
 
 fun provideSqlDriver(app: Application): SqlDriver {
