+++ androidApp/src/main/java/com/example/emobilitychargingstations/android/ui/auto/screen/ChargingMapScreen.kt
@@ -4,19 +4,26 @@ import android.graphics.Bitmap
 import android.location.Location
 import androidx.car.app.CarContext
 import androidx.car.app.OnScreenResultListener
-import androidx.car.app.model.*
+import androidx.car.app.model.Action
+import androidx.car.app.model.ActionStrip
+import androidx.car.app.model.CarColor
+import androidx.car.app.model.ItemList
+import androidx.car.app.model.ParkedOnlyOnClickListener
+import androidx.car.app.model.PlaceListMapTemplate
+import androidx.car.app.model.Row
+import androidx.car.app.model.Template
 import androidx.lifecycle.lifecycleScope
 import com.comsystoreply.emobilitychargingstations.android.BuildConfig
 import com.comsystoreply.emobilitychargingstations.android.R
 import com.example.emobilitychargingstations.android.ui.auto.BaseScreen
-import com.example.emobilitychargingstations.android.ui.utilities.buildRowWithPlace
-import com.example.emobilitychargingstations.android.ui.utilities.createCarIconFromBitmap
-import com.example.emobilitychargingstations.android.ui.utilities.getPlaceWithMarker
-import com.example.emobilitychargingstations.android.ui.utilities.getString
 import com.example.emobilitychargingstations.android.ui.utilities.AUTO_POI_MAP_SCREEN_MARKER
 import com.example.emobilitychargingstations.android.ui.utilities.LocationRequestStarter
 import com.example.emobilitychargingstations.android.ui.utilities.NAVIGATION_DISTANCE_VALUE_FOR_COMPLETION_IN_METERS
+import com.example.emobilitychargingstations.android.ui.utilities.buildRowWithPlace
+import com.example.emobilitychargingstations.android.ui.utilities.createCarIconFromBitmap
 import com.example.emobilitychargingstations.android.ui.utilities.getDrawableAsBitmap
+import com.example.emobilitychargingstations.android.ui.utilities.getPlaceWithMarker
+import com.example.emobilitychargingstations.android.ui.utilities.getString
 import com.example.emobilitychargingstations.android.ui.utilities.getTitleAsSpannableStringAndAddDistance
 import com.example.emobilitychargingstations.android.ui.utilities.getTransparentCarColor
 import com.example.emobilitychargingstations.data.extensions.getLatitude
