+++ androidApp/src/main/java/com/example/emobilitychargingstations/android/ui/composables/reusables/ActivityViewModelComposable.kt
@@ -3,7 +3,7 @@ package com.example.emobilitychargingstations.android.ui.composables.reusables
 import androidx.activity.ComponentActivity
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
-import com.example.emobilitychargingstations.android.StationsViewModel
+import com.example.emobilitychargingstations.android.ui.viewmodels.StationsViewModel
 import org.koin.androidx.compose.getViewModel
 
 @Composable
